Release 1.1.0  "Attack of the Donkeys"

Release 1.1.0 "Attack of the Donkeys"

See merge request !364

+++ lib/rucio/db/migrate_repo/versions/2ba5229cb54c_add_submitted_at_to_requests_table.py
@@ -19,13 +19,15 @@ Create Date: 2015-04-14 15:56:22.772281
 revision = '2ba5229cb54c'
 down_revision = '22d887e4ec0a'
 
-from alembic import op
+from alembic import op, context
 import sqlalchemy as sa
 
 
 def upgrade():
-    op.add_column('requests', sa.Column('submitted_at', sa.DateTime()))
+    if context.get_context().dialect.name not in ('sqlite'):
+        op.add_column('requests', sa.Column('submitted_at', sa.DateTime()))
 
 
 def downgrade():
-    op.drop_column('requests', 'submitted_at')
+    if context.get_context().dialect.name not in ('sqlite'):
+        op.drop_column('requests', 'submitted_at')
